# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freelancer Platform API
  description: |
    Comprehensive API for a global freelance marketplace with secure payments, 
    verified profiles, milestone-based delivery, and dispute resolution.
  version: 1.0.0
  contact:
    name: NivixPe Team
    email: api@freelancerplatform.com
servers:
  - url: https://api.freelancerplatform.com/v1
    description: Production
  - url: https://staging-api.freelancerplatform.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contractId, revieweeId, rating]
              properties:
                contractId:
                  type: string
                  format: uuid
                revieweeId:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '201':
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
  /reviews/user/{userId}:
    get:
      tags: [Reviews]
      summary: Get reviews for user
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  # ==================== DISPUTES ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [FREELANCER, CLIENT, ADMIN]
        fullName:
          type: string
        avatarUrl:
          type: string
        profileData:
          type: object
        rating:
          type: number
        totalReviews:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        reviewerId:
          type: string
          format: uuid
        revieweeId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        reviewer:
          $ref: '#/components/schemas/User'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...